fin test case

+++ simulation/Grid/Case.cpp
@@ -127,12 +127,12 @@ void moveA(Ant *a, Case *c1, Case *c2){
     c2->addAnt(a);
 
 }
-int norme(Coordonate c){
-    return sqrt(c.getLig() * c.getLig() + c.getCol() *c.getCol());
+float norme(Coordonate c){
+    return sqrt(float(c.getLig() * c.getLig() + c.getCol() *c.getCol()));
 }
 // a modif
-int distancePoint(Coordonate c1,  Coordonate c2){
-    return sqrt((c1.getLig()-c2.getLig()) * (c1.getLig()-c2.getLig()) + (c1.getCol()-c2.getCol()) *(c1.getCol()-c2.getCol()));
+float distancePoint(Coordonate c1,  Coordonate c2){
+    return sqrt(float((c1.getLig()-c2.getLig()) * (c1.getLig()-c2.getLig()) + (c1.getCol()-c2.getCol()) *(c1.getCol()-c2.getCol())));
 }
 bool closerN(Case *c1, Case *c2,Case *nest){
     if(nest->hasNest() == false) throw std::string("you need to have a nest on this case");
