+++ simulation/Grid/Case.hpp
@@ -48,8 +48,8 @@ class Case{
 
 };
 void moveA(Ant *a, Case *c1, Case *c2) ;
-int norme(Coordonate c);
-int distancePoint(Coordonate c1,  Coordonate c2);
+float norme(Coordonate c);
+float distancePoint(Coordonate c1,  Coordonate c2);
 bool closerN(Case *c1, Case *c2,Case *nest);
